basic quicksort algorithm
1.1. Naive Implementation of Quicksort:
To warm up, you are expected to implement the basic QuickSort sorting algorithm without
any optimization. The goal of this part is to ensure that you have a solid foundation of
the core principles of the QuickSort algorithm. Your code should:
1. Choose a pivot element from the array (for this part, use the last element as your
pivot).
2. Partition the array into two subarrays around the pivot. Elements on the left hand
side subarray should be smaller in value than the elements on the right hand side
subarray. Your algorithm is only expected to sort in ascending order.
3. Recursively apply QuickSort on the subarrays until you reach the base case.